+++ example/storybook/src/components/Overlay/Menu/Menu.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Wrapper from '../../Wrapper';
-import { Button, Menu, MenuIcon } from '../../../ui-components';
+import { Button, Menu, MenuIcon, Text } from '../../../ui-components';
 
 export const MenuStory = ({ placement }: any) => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -45,4 +45,4 @@ export const MenuStory = ({ placement }: any) => {
   );
 };
 
-export { Menu, Button, MenuIcon };
+export { Menu, Button, MenuIcon, Text };