+++ example/storybook/src/ui-components/Fab/styled-components/Root.tsx
@@ -76,7 +76,7 @@ export default styled(
         bg: '$primary500',
       },
       ':active': {
-        bg: '$prinary600',
+        bg: '$primary600',
       },
       ':disabled': {
         opacity: 0.4,