+++ packages/popper/src/PopperContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from '@gluestack-ui/utils';
-export const [PopperProvider, usePopperContext] =
-  createContext<any>('PopperContext');